Move RecorderWorkEntry in internal_model
