+++ catalogue_pipeline/recorder/src/main/scala/uk/ac/wellcome/platform/recorder/modules/RecorderWorkEntryKeyPrefixGeneratorModule.scala
@@ -1,10 +1,9 @@
 package uk.ac.wellcome.platform.recorder.modules
 
 import javax.inject.Singleton
-
 import com.google.inject.Provides
 import com.twitter.inject.TwitterModule
-import uk.ac.wellcome.platform.recorder.models.RecorderWorkEntry
+import uk.ac.wellcome.models.recorder.internal.RecorderWorkEntry
 import uk.ac.wellcome.storage.s3.{KeyPrefixGenerator, SourcedKeyPrefixGenerator}
 
 object RecorderWorkEntryKeyPrefixGeneratorModule extends TwitterModule {
